Refactor: refactor album title list component
