+++ src/components/mainComponents/AlbumTitles/AlbumTitlesStyle.css
